+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- FinitePoint: Use `ReadOnlySpan<char>` for `string` parameters in ctor and methods.
+
 ## [0.10.1] - 2023-05-08
 ### Fixed
 - Fixed BigIntCalculator's Equals method to avoid timing attacks. The slow equal implementation is used now.